pyproject.toml
[tool.poetry]
name = "recruiter-response-cli"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "recruiter_response_cli"}]

[tool.poetry.dependencies]
python = "^3.9"
typer = "^0.9.0"
rich = "^13.3.5"
pyyaml = "^6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
recruiter-response = "recruiter_response_cli.main:main"